Insert method for Part, carsystem. getMemberByIdAndPassword (for login)

+++ webshop/src/main/java/com/mycompany/carshop/model/beans/Member.java
@@ -23,20 +23,28 @@ public class Member {
     @GeneratedValue
     @Column(name = "id_member")
     private int memberId;
+
     @Column(name = "first_name")
     private String firstName;
+
     @Column(name = "last_name")
     private String lastName;
+
     @Column(name = "address")
     private String address;
+
     @Column(name = "phone")
     private String phone;
+
     @Column(name = "email")
     private String email;
+
     @Column(name = "money")
-    private double money;
+    private float money;
+
     @Column(name = "password")
     private String password;
+
     @ManyToOne
     @JoinColumn(name = "id_role")
     private Role role;
@@ -99,11 +107,11 @@ public class Member {
         this.email = email;
     }
 
-    public double getMoney() {
+    public float getMoney() {
         return money;
     }
 
-    public void setMoney(double money) {
+    public void setMoney(float money) {
         this.money = money;
     }
